Change targetFPS in Core.h from int to Uint64

+++ BeatEmLeft-Engine/BeatEmLeft-Engine/Core.cpp
@@ -19,7 +19,7 @@ Core::~Core()
 	appCleanUp();
 }
 
-int Core::getTargetFPS()
+Uint64 Core::getTargetFPS()
 {
 	return targetFPS;
 }
